Reuse the old preview buffer/window such that display-buffer behaves (#160)
* Reuse the old preview buffer/window such that display-buffer behaves

Before, the buffer is killed and then recreated. This causes wrong
behavior for display-buffer when display-buffer-reuse-window and
display-buffer-same-window is used: the first one fails so the same
window is used where the diagram source is situated.

Instead, reuse the buffer and its window, so don't kill the buffer,
erase it instead.

* Fix subsequent previews

image-mode marks buffer parts as read-only, essentially disallowing (erase-buffer).